requirements.txt
fastapi0==0.115.
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.9.0
pytesseract==0.3.10
Pillow==10.4.0
pdfplumber==0.11.2
python-docx==1.1.2
mysql-connector-python
